Ratings

+++ app/forms/comment.py
@@ -1,8 +1,8 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField
+from wtforms import StringField, IntegerField, DecimalField
 from wtforms.validators import InputRequired
 
 class NewCommentForm(FlaskForm):
   comment = StringField('comment', validators=[InputRequired(message='Please enter a comment')])
-  rating = IntegerField('rating')
+  rating = DecimalField('rating')
   profileId = IntegerField('profileId')
